fix: DB errors make crash bill's page
there are errors in the DB : some bills have no date, and some other bills have weird dates ( day-month-year instead of year-month-day). So it makes crash the formatDate function which is called to build the bill's table for generic user.
Database seems to be shared between all student working on the project, so if someone modifies the newbills validation procedure on his fork, he can upload invalid bills
Fix is made by adding a condition to return a valid (random) date if the input date is invalid